Splittet methods to files for pep8 and pylint conformity.
