+++ MobSF/utils.py
@@ -327,3 +327,16 @@ def isDirExists(dir_path):
 
 def genRandom():
     return ''.join([random.SystemRandom().choice('abcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*(-_=+)') for i in range(50)])
+
+
+def zipdir(path, zip_file):
+    """Zip a directory."""
+    try:
+        print "[INFO] Zipping"
+        # pylint: disable=unused-variable
+        # Needed by os.walk
+        for root, sub_dir, files in os.walk(path):
+            for file_name in files:
+                zip_file.write(os.path.join(root, file_name))
+    except:
+        PrintException("[ERROR] Zipping")
