+++ StaticAnalyzer/views/android/dvm_permissions.py
